requirements.txt
django
django-rest-framework
ipdb
django-solo
django-summernote
Pillow
